+++ src/util/parseDescription.ts
@@ -1,67 +1,5 @@
 import Description from '../structure/Description';
-
-type ColorCode =
-	| 'black'
-	| 'dark_blue'
-	| 'dark_green'
-	| 'dark_aqua'
-	| 'dark_red'
-	| 'dark_purple'
-	| 'gold'
-	| 'gray'
-	| 'dark_gray'
-	| 'blue'
-	| 'green'
-	| 'aqua'
-	| 'red'
-	| 'light_purple'
-	| 'yellow'
-	| 'white';
-
-type FormatCode =
-	| 'obfuscated'
-	| 'bold'
-	| 'strikethrough'
-	| 'underline'
-	| 'italic'
-	| 'reset';
-
-interface Chat {
-	text?: string,
-	bold?: 'true' | 'false',
-	italic?: 'true' | 'false',
-	underlined?: 'true' | 'false',
-	strikethrough?: 'true' | 'false',
-	obfuscated?: 'true' | 'false',
-	color?: ColorCode | FormatCode,
-	extra?: Chat[]
-}
-
-interface ModInfo {
-	type: string,
-	modList: {
-		modid: string,
-		version: string
-	}[]
-}
-
-interface RawResponse {
-	version?: {
-		name?: string,
-		protocol?: number
-	},
-	players?: {
-		max?: number,
-		online?: number,
-		sample?: {
-			name: string,
-			id: string
-		}[]
-	},
-	description?: Chat | string,
-	favicon?: string,
-	modinfo?: ModInfo
-}
+import { Chat } from '../model/Chat';
 
 const colorCodes = {
 	black: '0',
@@ -126,5 +64,4 @@ function parseDescription(description: Chat | string): Description {
 	return new Description(result);
 }
 
-export { RawResponse, Chat, ColorCode, FormatCode, ModInfo };
 export default parseDescription;
